backend/worker/requirements.txt
google-cloud-pubsub
diffusers
torch
transformers
accelerate
google-cloud-secret-manager